Add module to add filtering columns
